refactor version specfic io
