+++ src/powerfactory_tools/utils/__init__.py
@@ -2,8 +2,12 @@
 # :copyright: Copyright (c) Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden, 2022-2024.
 # :license: BSD 3-Clause
 
-from powerfactory_tools.versions.pf2024.utils.io import ExportHandler
+from powerfactory_tools.utils.io import BaseExportHandler
+from powerfactory_tools.utils.io import FileType
+from powerfactory_tools.utils.io import ImportHandler
 
 __all__ = [
-    "ExportHandler",
+    "BaseExportHandler",
+    "FileType",
+    "ImportHandler",
 ]
